CT-122 블로그 리스트 페이지 리팩토링
